allow delete patient, finish id int -> long migration

+++ src/main/java/de/adesso/termacare/database/repo/AddressRepo.java
@@ -3,7 +3,7 @@ package de.adesso.termacare.database.repo;
 import de.adesso.termacare.database.construct.AbstractRepo;
 import de.adesso.termacare.data.entity.Address;
 
-public class AddressRepo extends AbstractRepo<Address, Integer> {
+public class AddressRepo extends AbstractRepo<Address, Long> {
 
 	public AddressRepo(){
 		super(Address.class);
