+++ src/main/java/de/adesso/termacare/database/repo/MedicationRepo.java
@@ -3,7 +3,7 @@ package de.adesso.termacare.database.repo;
 import de.adesso.termacare.database.construct.AbstractRepo;
 import de.adesso.termacare.data.entity.Medication;
 
-public class MedicationRepo extends AbstractRepo<Medication, Integer>{
+public class MedicationRepo extends AbstractRepo<Medication, Long>{
 
 	public MedicationRepo(){
 		super(Medication.class);
